Create static-assets module to generate scripts and stylesheets
- Replace createScripts and createStylesheets with static-assets/
  - static-assets/dev provides assets for development
  - static-assets/prod provides assets for production
- The logic for conditionally importing the correct assets sits within
static-assets/index.js
