--- server/render/createStylesheets.js
@@ -1,8 +1,0 @@
-export const createStylesheets = () => ({
-  if (__PRODUCTION__) {
-    const assets = require('../../public/assets/manifest.json');
-    return `<link rel="stylesheet" href="/assets/${assets['styles/main.css']} />`
-  }
-
-});
-
