+++ server/render/static-assets/index.js
@@ -1,0 +1,4 @@
+const createStaticAssets = __PRODUCTION__ ? require('./prod') : require('./dev');
+
+export default createStaticAssets;
+
